requirements.txt
discord
git+https://github.com/clerie/mcuuid
loguru
dblpy